Move rhythm to utils
